requirements.txt
fastapi
uvicorn
python-dotenv
SQLAlchemy
psycopg2-binary
bcrypt==4.0.1
passlib[bcrypt]
python-jose[cryptography]
python-multipart
cryptography
